+++ cms_support/__init__.py
@@ -1,0 +1,7 @@
+# coding=utf-8
+
+from cms_support.utils.constants import Constants
+
+__version__ = Constants.VERSION
+__author__ = Constants.AUTHOR
+__credits__ = Constants.CREDIT
