+++ cms_support/transfers/transfers.py
@@ -11,86 +11,6 @@ import pandas as pd
 from utils.query_utils import AbstractQueries, get_lfn_and_short_pfn, group_data, Time, timestamp_to_human_utc
 import xlsxwriter
 
-"""
-######################## monit_prod_cms_rucio_enr*  ########################
-query = "data.name:T2_US_Florida AND data.status:ok AND metadata.path:hc15min"
-rucio_index = {"name": "monit_prod_cms_rucio_enr*", "id": "9732"}
-
-data.dst_rse 	T2_KR_KISTI
-data.src_rse 	T2_US_Wisconsin
-
-data.activity 	Production Output|User Subscriptions
-data.bytes 	111,809,924
-data.file_size 	111,809,924
-data.checksum_adler 	c88b8895
-data.file_id 	2867026540
-
--------------------- SUBMITTED -------------------------------
-data.event_type     transfer-submitted
-data.state 	SUBMITTED
-
-data.external_host 	https://fts3.cern.ch:8446
-data.name 	/store/mc/RunIIAutumn18NanoAODv7/VBFHHTo2G2ZTo2G.../x.root
-data.queued_at 	2020-10-06 12:58:45.839821
-data.request_type 	transfer
-data.scope 	cms
-data.vo 	cms
-metadata.producer 	cms_rucio
--------------------- FAILED -------------------------------
-data.event_type     transfer-failed
-data.state	unknown
-data.purged_reason 	DESTINATION [17] Destination file exists and overwrite is not enabled
-data.reason 	DESTINATION [17] Destination file exists and overwrite is not enabled
-
--------------------- OK -------------------------------
-data.event_type     transfer-done
-data.state	unknown
-
-data.transfer_endpoint 	https://fts3.cern.ch:8446
-data.src_url 	srm://gfe02.grid.hep.ph.ic.ac.uk:8443/srm/managerv2?SFN=/pnfs/hep.ph.ic.ac.uk/data/cms/.../tree_54.root
-data.dst_url 	srm://stormfe1.pi.infn.it:8444/srm/managerv2?SFN=/cms/.../tree_54.root
-data.source_se 	srm://dcache-se-cms.desy.de
-data.dest_se 	gsiftp://gridftp.accre.vanderbilt.edu
-data.name 	/store/.../tree_54.root
-"""
-"""
-##################################### monit_prod_fts_raw_* ###########################
-
-query = "data.vo:cms AND data.dest_se:/.*se.cis.gov.pl.*/ AND data.job_state:FAILED"
-fts_index = {"name": "monit_prod_fts_raw_*", "id": "9233"}
-
-query = "data.dst_url:/.*storm-fe-cms.cr.cnaf.infn.it.*/ AND data.file_state:FAILED AND data.reason:/.*CHECKSUM MISMATCH.*/" 
-
-class TransferAttributes:
-    attr_destination = "data.dst_url:"
-    attr_source = "data.src_url:"
-    attr_status = "data.file_state:"
-    attr_reason = "data.reason:"
-
-
-class Status:
-    error = "FAILED"
-    active = "ACTIVE"
-    ready = "READY"
-    submitted = "SUBMITTED"
-    finished = "FINISHED"
-
-def join_attribute_and_expression(self, attr, expression):
-    return CteFTS.AND + attr + CteFTS.EXPRESSION.format(CteFTS.INCLUDED.format(expression))
-
-def get_query(self, dst_url="", dst_protocol="", src_url="", src_protocol="", status="", reason=""):
-    raw_query = "data.vo:cms"
-    if dst_url:
-        raw_query += self.join_attribute_and_expression(self.attr_destination, dst_url)
-    if status:
-        raw_query += CteFTS.AND + self.attr_status + status
-
-    clean_str_query = get_str_lucene_query(self.index['name'], self.time_slot[0], self.time_slot[1], raw_query)
-    return clean_str_query
-
-https://monit-kibana.cern.ch/kibana/goto/31129642d019649a5fe8ee3c816677e6
-"""
-
 
 class Transfers(AbstractQueries, ABC):
     def __init__(self, time_class):
