+++ cms_support/utils/transfers_utils.py
@@ -1,9 +1,10 @@
 # coding=utf-8
+
 import time
 import re
 import pandas as pd
 import logging
-from tools.utils.constants import CteRucio
+from cms_support.utils.constants import CteRucio
 import xlsxwriter
